Remove map cascade and multiclass support; restrict cascade to binary predictions
BARGAIN's precision/recall guarantees require binary (positive/negative) labels,
and the accuracy guarantee works well on binary too. Multiclass support (accuracy
with enum labels) added complexity without the same statistical grounding.

- Remove _execute_map_cascade and _enum_output_key_and_values from MapOperation
- Move cascade config field and CascadeMixin from MapOperation to FilterOperation
- Remove multiclass branch from proxy_predict (always compute P(positive))
- Remove multiclass rendering from TUI detail pane
- Keep accuracy guarantee (works for binary alongside precision/recall)
- Update test_cascade_core to use binary labels
